% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/better_acos.R
\name{better.acos}
\alias{better.acos}
\title{Calculate The Inverse Of Cosine}
\usage{
better.acos(hypotenuse, adjacent, oppisite)
}
\arguments{
\item{hypotenuse}{Length of the a triangles hypotenuse}

\item{adjacent}{Signed length of an angles adjacent side. The sign describes the direction the triangle is pointing in a
cartesian coordinate system}

\item{oppisite}{Signed length of an angles oppisite side. The sign describes the direction the triangle is pointing in a
cartesian coordinate system}
}
\value{
Angle between the x-axis and the hypotenuse of an arbitrary triangle
}
\description{
This function is a wrapper around the standard acos function, which returns the corresponding angle between 0 and pi for any
output of the cosine function. This wrapper is able to return an angle between 0 and 2pi by looking at the sides of a right
angled triangle. The triangle angle this function returns is placed at the origin of a cartesian coordinate system and spans
between the hypotenuse and the adjacent side. The sign of the opposite side enables this function to return angles between 0
and 2pi.
}
