% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/better_subtraction.R
\name{better.subtraction}
\alias{better.subtraction}
\title{Calculates the smallest distance between two modular numbers}
\usage{
better.subtraction(diff, base = 2 * pi)
}
\arguments{
\item{diff}{difference: (b mod base)-(a mod base)}

\item{base}{base of the modulus numbers a and b.}
}
\value{
signed difference of b and a with the smallest possible absolute value in modulus base
}
\description{
This function takes the difference b-a and returns the signed difference with the smallest absolute value in modulus base.
The function will compute b mod base and a mod base before subtracting.
}
