% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_version.R
\name{check.version}
\alias{check.version}
\title{Check The Version Of The Package}
\usage{
check.version(expected.version)
}
\arguments{
\item{expected.version}{The version (as string) of RHotStuff the script was designed to use.}
}
\description{
This function makes sure that the scripts your running is working with the correct version
of RHotStuff. If the version does not match the expectations, the function prompts a warning.
}
