% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GET_elabftw.R
\name{GET.elabftw}
\alias{GET.elabftw}
\title{Request Experimental Data From eLabFTW Via API}
\usage{
GET.elabftw(
  experiment.id,
  node.id = "#experimental-data",
  dec = ",",
  api.key = Sys.getenv("READ_ELABFTW_TOKEN"),
  url = Sys.getenv("ELABFTW_API_URL")
)
}
\arguments{
\item{experiment.id}{The eLabFTW id of the experiment that will be fetched. The id can be found in the url of the experiment.}

\item{node.id}{A css selector that uniquely identifies the html table that should be read.}

\item{dec}{The decimal separator used in the table. The default is the German number notation with a comma instead of a period!}

\item{api.key}{The eLabFT API-key. See the eLabFTW documentation for details. The default settings retrieve this value from
the enviourmental variable READ_ELABFTW_TOKEN. This variable can be set in the ~/.Renviron file.}

\item{url}{The url to the eLabFTW API. See the eLabFTW documentation for details. The default settings retrieve this value from
the enviourmental variable ELABFTW_API_URL. This variable can be set in the ~/.Renviron file.}
}
\value{
A dataframe with the content of the selected html table from the eLabFTW online labbook.
}
\description{
This function fetches experimental data from the online labbook eLabFTW as a data frame. The html table that will be returned
as data frame musst be marked with a html property, like an id or a class. The table will be selected with the help of this
unique id, class or similar.
}
