% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polaram_as_spectrum.R
\name{polaram.as.spectrum}
\alias{polaram.as.spectrum}
\title{Compute A Raman Spectrum From PolaRam Output

This function takes the output of the polaram simulate program and computes one possible raman spectrum. PolaRam is a program that simulates the
raman scattering process with respect to the polarisation of the light. For details see the \href{https://github.com/AlreadyTakenJonas/PolaRam}{GitHub-Repository}.
The spectrum will be computed as sum of loretz curves. Every peak of the spectrum will be modeled by a lorentz curve. This function makes the assuption, that
the spectrum was created with a hypothetical photo detector, that responds differently to light polarised along the x-axis and y-axis. The two scalars scaleX
and scaleY allow to influence the height of the signal, depending of the polarisation of the hypothetically detectable light. If both values are one, the light
will be not distinguished by its polarisation.}
\usage{
polaram.as.spectrum(
  x,
  stokes.S0,
  stokes.S1,
  peaks.wavenumber,
  gamma = 25,
  normalise = FALSE,
  scaleX = 2,
  scaleY = 1
)
}
\arguments{
\item{x}{A vector of wavnumbers. The function computes for every wavenumber one point of the spectrum.}

\item{stokes.S0}{The first stokes parameter of the scattered light. It may be a vector with one element for each peak.}

\item{stokes.S1}{The second stokes parameter of the scattered light. It may be a vector with one element for each peak.}

\item{peaks.wavenumber}{The wavenumber of the raman peak. It may be a vector with one element for each peak.}

\item{gamma}{A scalar describing the width of the peaks. It may be a vector with one element for each peak. If the vector gamma is shorter than the peaks.wavenumber
vector, gamma will be reapeated until the length of both vectors match. The FWHM is in terms of the peaks width gamma (γ) and the peaks wavenumber (ν):
FWHM = sqrt(ν^2 + νγ) - sqrt(ν^2 - νγ) .}

\item{normalise}{Boolean value. If true the spectrum will be divided by its maximum.}

\item{scaleX}{Multiplicator, which describes how sensitive the photo detector is along the x-axis.}

\item{scaleY}{Multiplicator, which describes how sensitive the photo detector is along the y-axis.}
}
\value{
A data frame containing a calculated raman spectrum.
}
\description{
Compute A Raman Spectrum From PolaRam Output

This function takes the output of the polaram simulate program and computes one possible raman spectrum. PolaRam is a program that simulates the
raman scattering process with respect to the polarisation of the light. For details see the \href{https://github.com/AlreadyTakenJonas/PolaRam}{GitHub-Repository}.
The spectrum will be computed as sum of loretz curves. Every peak of the spectrum will be modeled by a lorentz curve. This function makes the assuption, that
the spectrum was created with a hypothetical photo detector, that responds differently to light polarised along the x-axis and y-axis. The two scalars scaleX
and scaleY allow to influence the height of the signal, depending of the polarisation of the hypothetically detectable light. If both values are one, the light
will be not distinguished by its polarisation.
}
